+++ db/hooks.php
@@ -29,4 +29,8 @@ $callbacks = [
         'hook' => \core_ai\hook\after_ai_provider_form_hook::class,
         'callback' => \aiprovider_router\hook_listener::class . '::set_form_definition_for_aiprovider_router',
     ],
+    [
+        'hook' => \core_course\hook\before_course_deleted::class,
+        'callback' => \aiprovider_router\hook_listener::class . '::delete_keys_for_deleted_course',
+    ],
 ];